Add RouteDataFragment support
Route data is intended to be used when several (or many)
sites look and act the same but need to have some settings
made available to the Fragments that come after it.

Eg A settings file (in json form) is keyed by the host header
that returns json with an account id or key to be used to
retrieve the data for that account.